Code clean-up

+++ source/src/main/java/com/neilalexander/jnacl/crypto/curve25519.java
@@ -102,7 +102,7 @@ public class curve25519 {
 
     add(a, 0, a, 0, minuspp, 0);
 
-    int negative = (int) (-((a[aoffset + 31] >>> 7) & 1));
+    int negative = (-((a[aoffset + 31] >>> 7) & 1));
 
     for (int j = 0; j < 32; ++j) a[aoffset + j] ^= negative & (aorig[j] ^ a[aoffset + j]);
   }
@@ -213,7 +213,7 @@ public class curve25519 {
     int[] workp = work, sp = s, rp = r;
 
     for (int pos = 254; pos >= 0; --pos) {
-      int b = ((int) ((e[pos / 8] & 0xFF) >>> (pos & 7)));
+      int b = ((e[pos / 8] & 0xFF) >>> (pos & 7));
       b &= 1;
       select(xzmb, xzm1b, xzm, xzm1, b);
       add(a0, 0, xzmb, 0, xzmbp, 32);
