+++ source/src/main/java/com/neilalexander/jnacl/crypto/hsalsa20.java
@@ -35,10 +35,10 @@ public class hsalsa20 {
   }
 
   static int load_littleendian(byte[] x, int offset) {
-    return ((int) (x[offset]) & 0xff)
-        | ((((int) (x[offset + 1]) & 0xff)) << 8)
-        | ((((int) (x[offset + 2]) & 0xff)) << 16)
-        | ((((int) (x[offset + 3]) & 0xff)) << 24);
+    return ((x[offset]) & 0xff)
+        | ((((x[offset + 1]) & 0xff)) << 8)
+        | ((((x[offset + 2]) & 0xff)) << 16)
+        | ((((x[offset + 3]) & 0xff)) << 24);
   }
 
   static void store_littleendian(byte[] x, int offset, int u) {
