+++ source/src/main/java/com/neilalexander/jnacl/crypto/verify_16.java
@@ -33,8 +33,8 @@ public class verify_16 {
   public static int crypto_verify(byte[] x, int xoffset, byte[] y) {
     int differentbits = 0;
 
-    for (int i = 0; i < 15; i++) differentbits |= ((int) (x[xoffset + i] ^ y[i])) & 0xff;
+    for (int i = 0; i < 15; i++) differentbits |= (x[xoffset + i] ^ y[i]) & 0xff;
 
-    return (1 & (((int) differentbits - 1) >>> 8)) - 1;
+    return (1 & ((differentbits - 1) >>> 8)) - 1;
   }
 }
